Move source files into appropriate directories.
