+++ index.js
@@ -1,19 +1,19 @@
-const RHOCore = require('./RHOCore');
+const RHOCore = require('./src/RHOCore');
 module.exports.RHOCore = RHOCore;
 
 
-const RNode = require('./rnodeAPI.js');
+const RNode = require('./src/rnodeAPI.js');
 module.exports.RNode = RNode.RNode;
 
 
-const signing = require('./signing');
+const signing = require('./src/signing');
 module.exports.keyPair = signing.keyPair;
 module.exports.verify = signing.verify;
 module.exports.b2h = signing.b2h;
 module.exports.h2b = signing.h2b;
 
 
-const hashing = require('./hashing.js');
+const hashing = require('./src/hashing.js');
 module.exports.sha256Hash = hashing.sha256Hash;
 module.exports.keccak256Hash = hashing.keccak256Hash;
 module.exports.blake2b256Hash = hashing.blake2b256Hash;
