+++ package.json
@@ -7,7 +7,7 @@
     "testAll": "npm run check && npm run integrationTest",
     "check": "npm run test && npm run lint && npm run flow-check",
     "test": "node test/testRHOCore.js; node test/testRNode.js; node test/testSigning.js",
-    "integrationTest": "node test/testRNode.js --net && node test/liveRHOCoreTest.js && node integrationTest.js",
+    "integrationTest": "node test/testRNode.js --net && node test/liveRHOCoreTest.js && node test/integrationTest.js",
     "flow-check": "flow check --from emacs",
     "flow-status": "flow status --from emacs",
     "lint": "eslint *.js **/*.js --format unix --ignore-pattern interfaces/ --ignore-pattern node_modules/ --ignore-pattern protobuf/"
