+++ src/hashing.js
